+++ core/src/executive/mod.rs
@@ -10,6 +10,24 @@ mod internal_contract;
 #[cfg(test)]
 mod executive_tests;
 
+trait CollateralCheckResultToVmResult {
+    fn into_vm_result(self) -> Result<(), vmError>;
+}
+
+impl CollateralCheckResultToVmResult for CollateralCheckResult {
+    fn into_vm_result(self) -> Result<(), vmError> {
+        match self {
+            CollateralCheckResult::ExceedStorageLimit { .. } => {
+                Err(vmError::ExceedStorageLimit)
+            }
+            CollateralCheckResult::NotEnoughBalance { required, got } => {
+                Err(vmError::NotEnoughBalanceForStorage { required, got })
+            }
+            CollateralCheckResult::Valid => Ok(()),
+        }
+    }
+}
+
 pub use self::{
     executed::*,
     executive::{
@@ -20,3 +38,5 @@ pub use self::{
         InternalContractTrait, SolidityFunctionTrait,
     },
 };
+use crate::vm::Error as vmError;
+use cfx_state::CollateralCheckResult;
